Eliminating static pattern

+++ MessageManagementServer/app/src/main/java/state/State.java
@@ -1,60 +1,54 @@
 package state;
 
-import com.rabbitmq.client.ConnectionFactory;
-import system.Message;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import system.Config;
 import system.Credentials;
+import system.Message;
 import system.User;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
-import java.io.IOException;
-import java.sql.*;
+import java.sql.DriverManager;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.concurrent.TimeoutException;
 
 /**
  * Class that holds the state of the application.
  */
 @Singleton
 public class State {
-    private com.rabbitmq.client.Connection rabbitMqConnection;
+    private static final Logger log = LoggerFactory.getLogger(State.class);
     private java.sql.Connection dbConnection;
-    private HashMap<String, User> activeUsers;
-    private HashMap<String, Message> activeMessages;
+    private final HashMap<String, User> activeUsers;
+    private final HashMap<String, Message> activeMessages;
 
     /**
      * Constructor for the State class.
      */
     @Inject
     public State() {
-        try {
-            // Initialize RabbitMQ connection
-            ConnectionFactory factory = new ConnectionFactory();
-            factory.setHost(Config.RABBITMQ_URL);
-            factory.setUsername(Config.RABBITMQ_USERNAME);
-            factory.setPassword(Config.RABBITMQ_PASSWORD);
-            rabbitMqConnection = factory.newConnection();
-        } catch (IOException | TimeoutException e) {
-            System.out.println("Failed to connect to RabbitMQ " + e.getMessage());
-        }
+        if (Config.DB_ENABLED) {
+            try {
+                // Initialize database connection
+                dbConnection = DriverManager.getConnection(Config.DB_URL, Config.DB_USERNAME, Config.DB_PASSWORD);
 
-        try {
-            // Initialize database connection
-            dbConnection = DriverManager.getConnection(Config.DB_URL, Config.DB_USERNAME, Config.DB_PASSWORD);
-
-            // Create tables if they don't exist
-            Statement statement = dbConnection.createStatement();
-            statement.execute("CREATE DATABASE IF NOT EXISTS e3ms;");
-            statement.execute("USE e3ms;");
-            statement.execute("CREATE TABLE IF NOT EXISTS users (id VARCHAR(255), name VARCHAR(255), username VARCHAR(255), password VARCHAR(255));");
-            statement.execute("CREATE TABLE IF NOT EXISTS messages (id VARCHAR(255), sender VARCHAR(255), receiver VARCHAR(255), timestamp VARCHAR(255), message VARCHAR(255));");
-            loadState(dbConnection.createStatement());
-        } catch (SQLException e) {
-            System.out.println("Failed to connect to database " + e.getMessage());
+                // Create tables if they don't exist
+                Statement statement = dbConnection.createStatement();
+                statement.execute("CREATE DATABASE IF NOT EXISTS e3ms;");
+                statement.execute("USE e3ms;");
+                statement.execute("CREATE TABLE IF NOT EXISTS users (id VARCHAR(255), name VARCHAR(255), username VARCHAR(255), password VARCHAR(255)), identitykey VARCHAR(255), signedprekey VARCHAR(255), prekeysignature VARCHAR(255), otpkey1 VARCHAR(255), otpkey2 VARCHAR(255), otpkey3 VARCHAR(255), otpkey4 VARCHAR(255);");
+                statement.execute("CREATE TABLE IF NOT EXISTS messages (id VARCHAR(255), sender VARCHAR(255), receiver VARCHAR(255), timestamp VARCHAR(255), message VARCHAR(255));");
+                loadState(dbConnection.createStatement());
+            } catch (SQLException e) {
+                log.error("Failed to connect to database " + e.getMessage());
+            }
         }
 
         // Initialize active maps
@@ -69,31 +63,48 @@ public class State {
      * @throws SQLException If an error occurs while executing the queries.
      */
     private void loadState(Statement statement) throws SQLException {
-        ResultSet users = statement.executeQuery("SELECT * FROM users");
-        while (users.next()) {
-            User user = new User(
-                    new Credentials(
-                            users.getString("username"), users.getString("password")
-                    ),
-                    users.getString("name"),
-                    users.getString("name")
-            );
-            user.setId(users.getString("id"));
-            activeUsers.put(user.getId(), user);
-        }
+        if (Config.DB_ENABLED) {
+            ResultSet users = statement.executeQuery("SELECT * FROM users");
+            while (users.next()) {
+                User user = new User(
+                        new Credentials(
+                                users.getString("username"),
+                                users.getString("password"),
+                                users.getString("identityKey"),
+                                users.getString("signedPreKey"),
+                                users.getString("preKeySignature"),
+                                new ArrayList<>() {
+                                    {
+                                        add(users.getString("otpKey1"));
+                                        add(users.getString("otpKey2"));
+                                        add(users.getString("otpKey3"));
+                                        add(users.getString("otpKey4"));
+                                    }
+                                }
+                        ),
+                        users.getString("name"),
+                        users.getString("name")
+                );
+                user.setId(users.getString("id"));
+                activeUsers.put(user.getId(), user);
+            }
 
-        ResultSet messages = statement.executeQuery("SELECT * FROM messages");
-        while (messages.next()) {
-            Message message = new Message(
-                    new Message.MetaData(
-                            messages.getString("id"),
-                            messages.getString("sender"),
-                            List.of(messages.getString("receiver").split(",")),
-                            messages.getString("timestamp")
-                    ),
-                    new Message.MessageData(messages.getString("message"))
-            );
-            activeMessages.put(message.getMetaData().getId(), message);
+            ResultSet messages = statement.executeQuery("SELECT * FROM messages");
+            while (messages.next()) {
+                Message message = new Message(
+                        new Message.MetaData(
+                                messages.getString("id"),
+                                messages.getString("sender"),
+                                List.of(messages.getString("receiver").split(",")),
+                                messages.getString("timestamp")
+                        ),
+                        new Message.MessageData(messages.getString("message"))
+                );
+                activeMessages.put(message.getMetaData().getId(), message);
+            }
+        } else {
+            log.info("Database is disabled, skipping loading state");
+            clear();
         }
     }
 
@@ -104,13 +115,34 @@ public class State {
      * @throws SQLException If an error occurs while executing the query.
      */
     public void addUser(User user) throws SQLException {
-        String sql = "INSERT INTO users (id, name, username, password) VALUES (?, ?, ?, ?)";
-        PreparedStatement statement = dbConnection.prepareStatement(sql);
-        statement.setString(1, user.getId());
-        statement.setString(2, user.getName());
-        statement.setString(3, user.getCredentials().getUsername());
-        statement.setString(4, user.getCredentials().getPassword());
-        statement.executeUpdate();
+        if (Config.DB_ENABLED) {
+            String sql = "INSERT INTO users (id, name, username, password, identitykey, signedprekey, prekeysignature, otpkey1, otpkey2, otpkey3, otpkey4) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)";
+            PreparedStatement statement = dbConnection.prepareStatement(sql);
+            statement.setString(1, user.getId());
+            statement.setString(2, user.getName());
+            statement.setString(3, user.getCredentials().getUsername());
+            statement.setString(4, user.getCredentials().getPassword());
+            statement.setString(5, user.getCredentials().getIdentityKey());
+            statement.setString(6, user.getCredentials().getSignedPreKey());
+            statement.setString(7, user.getCredentials().getPreKeySignature());
+            statement.setString(
+                    8,
+                    user.getCredentials().getOneTimePreKeys().get(0) == null ? "" : user.getCredentials().getOneTimePreKeys().get(0)
+            );
+            statement.setString(
+                    9,
+                    user.getCredentials().getOneTimePreKeys().get(1) == null ? "" : user.getCredentials().getOneTimePreKeys().get(1)
+            );
+            statement.setString(
+                    10,
+                    user.getCredentials().getOneTimePreKeys().get(2) == null ? "" : user.getCredentials().getOneTimePreKeys().get(2)
+            );
+            statement.setString(
+                    11,
+                    user.getCredentials().getOneTimePreKeys().get(3) == null ? "" : user.getCredentials().getOneTimePreKeys().get(3)
+            );
+            statement.executeUpdate();
+        }
         activeUsers.put(user.getId(), user);
     }
 
@@ -121,24 +153,17 @@ public class State {
      * @throws SQLException If an error occurs while executing the query.
      */
     public void addMessage(Message message) throws SQLException {
-        String sql = "INSERT INTO messages (id, sender, receiver, timestamp, message) VALUES (?, ?, ?, ?, ?)";
-        PreparedStatement statement = dbConnection.prepareStatement(sql);
-        statement.setString(1, message.getMetaData().getId());
-        statement.setString(2, message.getMetaData().getSender());
-        statement.setString(3, String.join(",", message.getMetaData().getReceiver())); // Assuming receiver is a list of strings
-        statement.setString(4, message.getMetaData().getTimestamp());
-        statement.setString(5, message.getMessageData().getMessage());
-        statement.executeUpdate();
+        if (Config.DB_ENABLED) {
+            String sql = "INSERT INTO messages (id, sender, receiver, timestamp, message) VALUES (?, ?, ?, ?, ?)";
+            PreparedStatement statement = dbConnection.prepareStatement(sql);
+            statement.setString(1, message.getMetaData().getId());
+            statement.setString(2, message.getMetaData().getSender());
+            statement.setString(3, String.join(",", message.getMetaData().getReceiver())); // Assuming receiver is a list of strings
+            statement.setString(4, message.getMetaData().getTimestamp());
+            statement.setString(5, message.getMessageData().getMessage());
+            statement.executeUpdate();
+        }
         activeMessages.put(message.getMetaData().getId(), message);
-    }
-
-    /**
-     * Get the RabbitMQ connection.
-     *
-     * @return The RabbitMQ connection.
-     */
-    public com.rabbitmq.client.Connection getQueueConnection() {
-        return rabbitMqConnection;
     }
 
     /**
@@ -198,7 +223,7 @@ public class State {
      * @return A list of all the messages in the state.
      */
     public List<Message> getMessages() {
-        return new LinkedList<>(activeMessages.values());
+        return new ArrayList<>(activeMessages.values());
     }
 
     /**
@@ -207,7 +232,7 @@ public class State {
      * @return A list of all the users in the state.
      */
     public List<User> getUsers() {
-        return new LinkedList<>(activeUsers.values());
+        return new ArrayList<>(activeUsers.values());
     }
 
     /**
@@ -217,15 +242,17 @@ public class State {
      * @throws SQLException If an error occurs while executing the query.
      */
     public void removeUser(User user) throws SQLException {
-        String sql = "DELETE FROM users WHERE id = ?";
-        PreparedStatement statement = dbConnection.prepareStatement(sql);
-        statement.setString(1, user.getId());
-        statement.executeUpdate();
+        if (Config.DB_ENABLED) {
+            String sql = "DELETE FROM users WHERE id = ?";
+            PreparedStatement statement = dbConnection.prepareStatement(sql);
+            statement.setString(1, user.getId());
+            statement.executeUpdate();
 
-        sql = "DELETE FROM messages WHERE sender = ?";
-        statement = dbConnection.prepareStatement(sql);
-        statement.setString(1, user.getId());
-        statement.executeUpdate();
+            sql = "DELETE FROM messages WHERE sender = ?";
+            statement = dbConnection.prepareStatement(sql);
+            statement.setString(1, user.getId());
+            statement.executeUpdate();
+        }
 
         activeUsers.remove(user.getId());
     }
@@ -237,22 +264,24 @@ public class State {
      * @throws SQLException If an error occurs while executing the query.
      */
     public void removeMessage(Message message) throws SQLException {
-        String sql = "DELETE FROM messages WHERE id = ?";
-        PreparedStatement statement = dbConnection.prepareStatement(sql);
-        statement.setString(1, message.getMetaData().getId());
-        statement.executeUpdate();
+        if (Config.DB_ENABLED) {
+            String sql = "DELETE FROM messages WHERE id = ?";
+            PreparedStatement statement = dbConnection.prepareStatement(sql);
+            statement.setString(1, message.getMetaData().getId());
+            statement.executeUpdate();
+        }
         activeMessages.remove(message.getMetaData().getId());
     }
 
     /**
      * Close the connections.
      *
-     * @throws IOException  If an error occurs while closing the RabbitMQ connection.
      * @throws SQLException If an error occurs while closing the database connection.
      */
-    public void close() throws IOException, SQLException {
-        rabbitMqConnection.close();
-        dbConnection.close();
+    public void close() throws SQLException {
+        if (Config.DB_ENABLED) {
+            dbConnection.close();
+        }
     }
 
     /**
